Refactor component factory
